# Mac kext performance tracing