--- a/ProjFS.Mac/PrjFSKext/PrjFSKext/KauthHandler.cpp
+++ b/ProjFS.Mac/PrjFSKext/PrjFSKext/KauthHandler.cpp
@@ -13,6 +13,7 @@
 #include "Message.h"
 #include "Locks.hpp"
 #include "PrjFSProviderUserClient.hpp"
+#include "PerformanceTracing.hpp"
 
 // Function prototypes
 static int HandleVnodeOperation(
@@ -61,6 +62,7 @@ static bool ShouldHandleVnodeOpEvent(
     vfs_context_t context,
     const vnode_t vnode,
     kauth_action_t action,
+    ProfileSample& operationSample,
 
     // Out params:
     VirtualizationRootHandle* root,
@@ -252,6 +254,8 @@ static int HandleVnodeOperation(
     uintptr_t       arg3)
 {
     atomic_fetch_add(&s_numActiveKauthEvents, 1);
+
+    ProfileSample functionSample(Probe_VnodeOp);
     
     vfs_context_t context = reinterpret_cast<vfs_context_t>(arg0);
     vnode_t currentVnode =  reinterpret_cast<vnode_t>(arg1);
@@ -296,6 +300,7 @@ static int HandleVnodeOperation(
             context,
             currentVnode,
             action,
+            functionSample,
             &root,
             &vnodeType,
             &currentVnodeFileFlags,
@@ -343,6 +348,8 @@ static int HandleVnodeOperation(
             // Recursively expand directory on delete to ensure child placeholders are created before rename operations
             if (isDeleteAction || FileFlagsBitIsSet(currentVnodeFileFlags, FileFlags_IsEmpty))
             {
+                functionSample.SetProbe(Probe_VnodeOp_PopulatePlaceholderDirectory);
+
                 if (!TrySendRequestAndWaitForResponse(
                         root,
                         isDeleteAction ?
@@ -376,6 +383,8 @@ static int HandleVnodeOperation(
         {
             if (FileFlagsBitIsSet(currentVnodeFileFlags, FileFlags_IsEmpty))
             {
+                functionSample.SetProbe(Probe_VnodeOp_HydratePlaceholderFile);
+
                 if (!TrySendRequestAndWaitForResponse(
                         root,
                         MessageType_KtoU_HydrateFile,
@@ -416,6 +425,8 @@ static int HandleFileOpOperation(
 {
     atomic_fetch_add(&s_numActiveKauthEvents, 1);
     
+    ProfileSample functionSample(Probe_FileOp);
+
     vfs_context_t context = vfs_context_create(NULL);
     vnode_t currentVnodeFromPath = NULLVP;
 
@@ -567,6 +578,7 @@ static bool ShouldHandleVnodeOpEvent(
     vfs_context_t context,
     const vnode_t vnode,
     kauth_action_t action,
+    ProfileSample& operationSample,
 
     // Out params:
     VirtualizationRootHandle* root,
@@ -582,6 +594,7 @@ static bool ShouldHandleVnodeOpEvent(
     
     if (!VirtualizationRoot_VnodeIsOnAllowedFilesystem(vnode))
     {
+        operationSample.SetProbe(Probe_Op_EarlyOut);
         *kauthResult = KAUTH_RESULT_DEFER;
         return false;
     }
@@ -589,16 +602,22 @@ static bool ShouldHandleVnodeOpEvent(
     *vnodeType = vnode_vtype(vnode);
     if (ShouldIgnoreVnodeType(*vnodeType, vnode))
     {
+        operationSample.SetProbe(Probe_Op_EarlyOut);
         *kauthResult = KAUTH_RESULT_DEFER;
         return false;
     }
     
-    *vnodeFileFlags = ReadVNodeFileFlags(vnode, context);
+    {
+        ProfileSample readflags(Probe_ReadFileFlags);
+        *vnodeFileFlags = ReadVNodeFileFlags(vnode, context);
+    }
+
     if (!FileFlagsBitIsSet(*vnodeFileFlags, FileFlags_IsInVirtualizationRoot))
     {
         // This vnode is not part of ANY virtualization root, so exit now before doing any more work.
         // This gives us a cheap way to avoid adding overhead to IO outside of a virtualization root.
         
+        operationSample.SetProbe(Probe_Op_NoVirtualizationRootFlag);
         *kauthResult = KAUTH_RESULT_DEFER;
         return false;
     }
@@ -618,14 +637,21 @@ static bool ShouldHandleVnodeOpEvent(
             // get called again, so we lose the opportunity to hydrate the file/directory and it will appear as though
             // it is missing its contents.
             
+            operationSample.SetProbe(Probe_Op_DenyCrawler);
             *kauthResult = KAUTH_RESULT_DENY;
             return false;
         }
+
+        operationSample.SetProbe(Probe_Op_EmptyFlag);
     }
     
+    operationSample.TakeSplitSample(Probe_Op_IdentifySplit);
+
     *vnodeFsidInode = Vnode_GetFsidAndInode(vnode, context);
     *root = VirtualizationRoot_FindForVnode(vnode, *vnodeFsidInode);
     
+    operationSample.TakeSplitSample(Probe_Op_VirtualizationRootFindSplit);
+
     if (RootHandle_ProviderTemporaryDirectory == *root)
     {
         *kauthResult = KAUTH_RESULT_DEFER;
@@ -650,6 +676,7 @@ static bool ShouldHandleVnodeOpEvent(
     // If the calling process is the provider, we must exit right away to avoid deadlocks
     if (VirtualizationRoot_PIDMatchesProvider(*root, *pid))
     {
+        operationSample.SetProbe(Probe_Op_Provider);
         *kauthResult = KAUTH_RESULT_DEFER;
         return false;
     }