--- a/ProjFS.Mac/PrjFSKext/PrjFSKext/PrjFSLogUserClient.hpp
+++ b/ProjFS.Mac/PrjFSKext/PrjFSKext/PrjFSLogUserClient.hpp
@@ -26,5 +26,18 @@ public:
     virtual IOReturn clientMemoryForType(UInt32 type, IOOptionBits* options, IOMemoryDescriptor** memory) override;
     virtual IOReturn registerNotificationPort(mach_port_t port, UInt32 type, io_user_reference_t refCon) override;
     
+    virtual IOReturn externalMethod(
+        uint32_t selector,
+        IOExternalMethodArguments* arguments,
+        IOExternalMethodDispatch* dispatch = nullptr,
+        OSObject* target = nullptr,
+        void* reference = nullptr) override;
+
+    
+    static IOReturn fetchProfilingData(
+        OSObject* target,
+        void* reference,
+        IOExternalMethodArguments* arguments);
+    
     void sendLogMessage(KextLog_MessageHeader* message, uint32_t size);
 };