… ProjFS.Mac/PrjFSKext/PrjFSKext/PrjFSProviderUserClient.hpp
… ProjFS.Mac/PrjFSKext/PrjFSKext/VirtualizationRoots.cpp
… ProjFS.Mac/PrjFSKext/public/PrjFSCommon.h
… ProjFS.Mac/PrjFSLib/PrjFSLib.xcodeproj/project.pbxproj
… ProjFS.Mac/PrjFSLib/prjfs-log/kext-perf-tracing.cpp
… ProjFS.Mac/PrjFSLib/prjfs-log/kext-perf-tracing.hpp
… ProjFS.Mac/PrjFSLib/prjfs-log/prjfs-log.cpp