Adds a simple mechanism for tracing the runtime of code paths in the kext.

This is an updated and reworked version of the kext performance tracing code I wrote ages ago. The [original PR (pre-github)](https://dev.azure.com/mseng/VSOnline/_git/GVFS/pullrequest/348922?_a=overview) was never merged due to the derived works licensing issue. This new PR no longer imports any external libraries.

Example output:

```
 0                                  VnodeOp    324663 [  324663] samples, total time:      2183224415 ns, mean:    6724.59 ns +/-   165772.00, min:     552 ns, max:   63021858 ns
 1                         VnodeOp_EarlyOut     14065 [   14065] samples, total time:          919921 ns, mean:      65.40 ns +/-      148.00, min:      10 ns, max:      10823 ns
 2         VnodeOp_NoVirtualizationRootFlag   1053931 [ 1053931] samples, total time:       324526160 ns, mean:     307.92 ns +/-      851.00, min:     131 ns, max:     506253 ns
 3                        VnodeOp_EmptyFlag         0 [       0] samples, total time:               0 ns, mean:       0.00 ns +/-        0.00
 4                      VnodeOp_DenyCrawler      6211 [    6211] samples, total time:         2723004 ns, mean:     438.42 ns +/-      199.00, min:     208 ns, max:       5000 ns
 5                          VnodeOp_Offline     13787 [   13787] samples, total time:       128897126 ns, mean:    9349.18 ns +/-    22885.00, min:    1476 ns, max:    1960524 ns
 6                         VnodeOp_Provider     33311 [   33311] samples, total time:       165406431 ns, mean:    4965.52 ns +/-     3066.00, min:     712 ns, max:     192446 ns
 7     VnodeOp_PopulatePlaceholderDirectory       405 [     405] samples, total time:      1679875615 ns, mean: 4147841.02 ns +/- 11614989.00, min:  229834 ns, max:  169771434 ns
 8           VnodeOp_HydratePlaceholderFile      3711 [    3711] samples, total time:      2797867855 ns, mean:  753939.06 ns +/-  9880228.00, min:  101931 ns, max:  599370787 ns
 9                    VnodeOp_IdentifySplit    375877 [  375877] samples, total time:       346696712 ns, mean:     922.37 ns +/-   153833.00, min:     193 ns, max:   63019327 ns
10      VnodeOp_VirtualizationRootFindSplit    375877 [  375877] samples, total time:      2087201093 ns, mean:    5552.88 ns +/-     9195.00, min:     299 ns, max:    1948646 ns
11                            ReadFileFlags   1436019 [ 1436019] samples, total time:       214778997 ns, mean:     149.57 ns +/-      599.00, min:      58 ns, max:     505953 ns
12                  VirtualizationRoot_Find    375877 [  375877] samples, total time:      2024552898 ns, mean:    5386.21 ns +/-     9150.00, min:     201 ns, max:    1948488 ns
13         VirtualizationRoot_FindIteration   1369038 [ 1369038] samples, total time:      1894929183 ns, mean:    1384.13 ns +/-     4698.00, min:      95 ns, max:    1945764 ns

```

The main functionality is built around a scoped time measurement class, and an enum of probe indices. Each index corresponds to a struct of aggregate data from samples. (total runtime, sum-of-squares for standard deviation calculation, minimum, maximum, number of samples)

The logging user client is extended by a method for extracting the data, and the sample user space logging tool pulls and outputs that data periodically.

The second commit places a bunch of probes throughout the kext, and the final 2 commits ensure that the instrumentation normally compiles down to nothing, and add a profiling build configuration where it's enabled.