@pmj 2018-10-01
I've updated this PR for latest master.